WhatsFinance/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
dateparser==1.2.0
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
iso8601==2.1.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
openai==1.59.7
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
pure_eval==0.2.3
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.1
pypika-tortoise==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
tornado==6.4.2
tortoise-orm==0.23.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
